working multiplayer match

+++ src/components/Providers/useLobby.js
@@ -1,6 +1,6 @@
-import firestore from '@react-native-firebase/firestore';
-import React, { useRef, useEffect, useState } from 'react';
-import { useFocusEffect } from '@react-navigation/native';
+import firestore from "@react-native-firebase/firestore";
+import React, { useRef, useState } from "react";
+import { useFocusEffect } from "@react-navigation/native";
 
 const useLobby = ({ lobbyId, userId }) => {
   const subscriber = useRef(() => {});
@@ -13,23 +13,29 @@ const useLobby = ({ lobbyId, userId }) => {
       if (lobbyId !== undefined) {
         subscriber.current();
         subscriber.current = firestore()
-          .collection('Lobbies')
+          .collection("Lobbies")
           .doc(lobbyId)
-          .onSnapshot((documentSnapshot) => {
-            setLobby(documentSnapshot.data());
-            lobbyRef.current = documentSnapshot.data();
-          }, (error) => console.log(error));
+          .onSnapshot(
+            (documentSnapshot) => {
+              setLobby(documentSnapshot.data());
+              lobbyRef.current = documentSnapshot.data();
+            },
+            (error) => console.log(error)
+          );
       } else {
         subscriber.current();
         subscriber.current = firestore()
-          .collection('Lobbies')
-          .onSnapshot((querySnapshot) => {
-            const lobbyArray = [];
-            querySnapshot.forEach(
-              (snapLobby) => { lobbyArray.push({ id: snapLobby.id, ...snapLobby.data() }); },
-            );
-            setLobbies(lobbyArray);
-          }, (error) => console.log(error));
+          .collection("Lobbies")
+          .onSnapshot(
+            (querySnapshot) => {
+              const lobbyArray = [];
+              querySnapshot.forEach((snapLobby) => {
+                lobbyArray.push({ id: snapLobby.id, ...snapLobby.data() });
+              });
+              setLobbies(lobbyArray);
+            },
+            (error) => console.log(error)
+          );
       }
 
       // Stop listening for updates when no longer required
@@ -37,29 +43,43 @@ const useLobby = ({ lobbyId, userId }) => {
         subscriber.current();
         if (lobbyId) {
           if (lobbyRef.current?.host?.uid === userId) {
-            firestore()
-              .collection('Lobbies')
-              .doc(lobbyId)
-              .delete();
+            firestore().collection("Lobbies").doc(lobbyId).delete();
           } else {
-            firestore().collection('Lobbies').doc(lobbyId).update({ [`players.${userId}`]: firestore.FieldValue.delete() })
+            firestore()
+              .collection("Lobbies")
+              .doc(lobbyId)
+              .update({ [`players.${userId}`]: firestore.FieldValue.delete() })
               .catch(() => {});
           }
         }
       };
-    }, [lobbyId, userId]),
+    }, [lobbyId, userId])
   );
 
-  const createLobby = (data) => firestore()
-    .collection('Lobbies')
-    .add(data);
+  const createLobby = (data) => firestore().collection("Lobbies").add(data);
 
-  const joinLobby = (id, user) => firestore().collection('Lobbies').doc(id).update({ [`players.${user.uid}`]: { name: user.email, ready: false } });
+  const joinLobby = (id, user) =>
+    firestore()
+      .collection("Lobbies")
+      .doc(id)
+      .update({ [`players.${user.uid}`]: { name: user.email, ready: false } });
 
-  const readyUp = () => firestore().collection('Lobbies').doc(lobbyId).update({ [`players.${userId}.ready`]: true });
+  const startMatch = (id) =>
+    firestore().collection("Lobbies").doc(lobbyId).update({ matchId: id });
+
+  const readyUp = () =>
+    firestore()
+      .collection("Lobbies")
+      .doc(lobbyId)
+      .update({ [`players.${userId}.ready`]: true });
 
   return {
-    createLobby, lobbies, lobby, joinLobby, readyUp,
+    createLobby,
+    lobbies,
+    lobby,
+    joinLobby,
+    readyUp,
+    startMatch,
   };
 };
